backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.1
python-dotenv==1.0.0
httpx==0.26.0
stellar-sdk==9.0.0
PyJWT==2.8.0
